+++ Assets/Scripts/Enemy/Enemy.cs
@@ -1,0 +1,129 @@
+﻿using UnityEngine;
+using UnityEngine.AI;
+using System.Collections;
+
+[RequireComponent(typeof(NavMeshAgent))]
+public class Enemy : LivingEntity {
+
+    public enum State { Idle, Chasing, Attacking };
+    State currentState;
+
+    public float refreshRate = 0.25f;
+    public float damage = 1;
+    public ParticleSystem deathEffect;
+
+    NavMeshAgent pathfinder;
+    Transform target;
+    LivingEntity targetEntity;
+    Material skinMaterial;
+
+    Color originalColour;
+
+    float attackDistanceThreshold = .5f;
+    float timeBetweenAttacks = 1;
+
+    float nextAttackTime;
+    float myCollisionRadius;
+    float targetCollisionRadius;
+
+    bool hasTarget;
+
+    protected override void Start() {
+        base.Start();
+        pathfinder = GetComponent<NavMeshAgent>();
+        skinMaterial = GetComponent<Renderer>().material;
+        originalColour = skinMaterial.color;
+
+        // at start it chases the player
+        if (GameObject.FindGameObjectWithTag("Player") != null) {
+            currentState = State.Chasing;
+            hasTarget = true;
+
+            target = GameObject.FindGameObjectWithTag("Player").transform;
+            targetEntity = target.GetComponent<LivingEntity>();
+
+            myCollisionRadius = GetComponent<CapsuleCollider>().radius;
+            targetCollisionRadius = target.GetComponent<CapsuleCollider>().radius;
+
+            StartCoroutine(UpdatePath());
+        }
+    }
+
+    // if player si death wait
+    void OnTargetDeath() {
+        hasTarget = false;
+        currentState = State.Idle;
+    }
+
+    public override void TakeHit(float damage, Vector3 hitPoint, Vector3 hitDirection) {
+        if (damage >= health) {
+            Destroy(Instantiate(deathEffect.gameObject, hitPoint, Quaternion.FromToRotation(Vector3.forward, hitDirection)), deathEffect.main.startLifetime.constant);
+        }
+        base.TakeHit(damage, hitPoint, hitDirection);
+    }
+
+    void Update() {
+        // if player is still alive
+        if (hasTarget) {
+            // wait between attacks
+            if (Time.time > nextAttackTime) {
+                float sqrDstToTarget = (target.position - transform.position).sqrMagnitude;
+                if (sqrDstToTarget < Mathf.Pow(attackDistanceThreshold + myCollisionRadius + targetCollisionRadius, 2)) {
+                    nextAttackTime = Time.time + timeBetweenAttacks;
+                    StartCoroutine(Attack());
+                }
+
+            }
+        }
+    }
+
+    IEnumerator Attack() {
+
+        // attack the player
+        currentState = State.Attacking;
+        pathfinder.enabled = false;
+
+        Vector3 originalPosition = transform.position;
+        Vector3 dirToTarget = (target.position - transform.position).normalized;
+        Vector3 attackPosition = target.position - dirToTarget * (myCollisionRadius);
+
+        float attackSpeed = 3;
+        float percent = 0;
+
+        skinMaterial.color = Color.red;
+        bool hasAppliedDamage = false;
+
+        // move the enemy near to simulate animation
+        while (percent <= 1) {
+
+            if (percent >= .5f && !hasAppliedDamage) {
+                hasAppliedDamage = true;
+                targetEntity.TakeDamage(damage);
+            }
+
+            percent += Time.deltaTime * attackSpeed;
+            float interpolation = (-Mathf.Pow(percent, 2) + percent) * 4;
+            transform.position = Vector3.Lerp(originalPosition, attackPosition, interpolation);
+
+            yield return null;
+        }
+
+        skinMaterial.color = originalColour;
+        currentState = State.Chasing;
+        pathfinder.enabled = true;
+    }
+
+    IEnumerator UpdatePath() {
+        while (hasTarget) {
+            // move the enemy
+            if (currentState == State.Chasing) {
+                Vector3 dirToTarget = (target.position - transform.position).normalized;
+                Vector3 targetPosition = target.position - dirToTarget * (myCollisionRadius + targetCollisionRadius + attackDistanceThreshold / 2);
+                if (!dead) {
+                    pathfinder.SetDestination(targetPosition);
+                }
+            }
+            yield return new WaitForSeconds(refreshRate);
+        }
+    }
+}