+++ Assets/Scripts/Player/Player.cs
@@ -1,0 +1,42 @@
+﻿using UnityEngine;
+
+[RequireComponent(typeof(PlayerController))]
+[RequireComponent(typeof(GunController))]
+public class Player : LivingEntity {
+
+    public float moveSpeed = 5;
+
+    Camera viewCamera;
+    PlayerController controller;
+    GunController gunController;
+
+    protected override void Start() {
+        base.Start();
+        controller = GetComponent<PlayerController>();
+        gunController = GetComponent<GunController>();
+        viewCamera = Camera.main;
+    }
+
+    void Update() {
+        // Movement input
+        Vector3 moveInput = new Vector3(Input.GetAxisRaw("Horizontal"), 0, Input.GetAxisRaw("Vertical"));
+        Vector3 moveVelocity = moveInput.normalized * moveSpeed;
+        controller.Move(moveVelocity);
+
+        // Look input
+        Ray ray = viewCamera.ScreenPointToRay(Input.mousePosition);
+        Plane groundPlane = new Plane(Vector3.up, Vector3.zero);
+        float rayDistance;
+
+        // interpole the plane with the point where the mouse is
+        if (groundPlane.Raycast(ray, out rayDistance)) {
+            Vector3 point = ray.GetPoint(rayDistance);
+            controller.LookAt(point);
+        }
+
+        // Weapon input for shoot the projectile
+        if (Input.GetMouseButton(0)) {
+            gunController.Shoot();
+        }
+    }
+}