+++ CHANGELOG.md
@@ -5,7 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.1] - 2018-06-09
+## [0.2.0] - 2018-20-09
+
+### Added
+
+- Player movement
+- Shooting system
+- Enemy behaviour
+- Map generation
+- Spawning system
+
+### Fixed
+
+- Fixed CHANGELOG.md errors
+
+## [0.1.2] - 2018-19-09
 
 ### Changed